+++ tests/integration/test_multi_accounts.py
@@ -1,15 +1,4 @@
-import pytest
-
-from localstack.testing.pytest.fixtures import _client
 from localstack.utils.strings import short_uid
-
-
-@pytest.fixture
-def client_factory():
-    def _client_factory(service: str, aws_access_key_id: str, region_name: str = "eu-central-1"):
-        return _client(service, region_name=region_name, aws_access_key_id=aws_access_key_id)
-
-    yield _client_factory
 
 
 class TestMultiAccounts:
